#143 Bot Einbindung

See merge request course/gadsen!29

+++ core/src/com/gats/simulation/TileAction.java
@@ -1,7 +1,5 @@
 package com.gats.simulation;
 
-import com.badlogic.gdx.math.Vector2;
-
 /**
  * Oberklasse für alle {@link Action Ereignisse}, die sich auf eine {@link Tile Box} beziehen
  */
@@ -9,7 +7,7 @@ abstract class TileAction extends Action{
 
 
     private IntVector2 pos;
-    protected TileAction(IntVector2 tilePos, long delay) {
+    protected TileAction(IntVector2 tilePos, float delay) {
         super(delay);
         this.pos = tilePos;
     }
